requirements.txt
Django==3.0.8
django-bootstrap4==2.2.0
django-crispy-forms==1.9.1
django-recaptcha==2.0.6